requirements.txt
aiogram>=3.4.0
motor>=3.3.2
pymongo>=4.6.1
pdfminer.six>=20231228
python-docx>=1.1.2
python-dotenv>=1.0.1
pydantic>=2.7.0
sentry-sdk>=2.9.0
prometheus-client>=0.20.0
fastapi>=0.115.0
uvicorn>=0.30.6
httpx>=0.27.2
